+++ src/NTwitch.Core/Entities/IEntity.cs
@@ -3,6 +3,6 @@
     public interface IEntity
     {
         ITwitchClient Client { get; }
-        ulong Id { get; }
+        uint Id { get; }
     }
 }
